slightly refactor blocking CLI code
