+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod os_backend;
 pub mod blocking {
-    pub mod manage_websites;
+    pub mod block_websites;
 }
 pub mod config {
